create a gpu base module
